55. Instalar shopping cart
utilize mejor la utileria  darryldecode/laravelshoppingcart
https://github.com/darryldecode/laravelshoppingcart

+++ database/migrations/2022_03_01_024453_create_products_table.php
@@ -21,7 +21,9 @@ return new class extends Migration
             $table->string('name');
 			$table->string('slug');
 			$table->text('description');
-			$table->float('price');
+
+			// $table->float('price');
+            $table->decimal('price',9,2);
 
 			$table->unsignedBigInteger('subcategory_id');
             $table->foreign('subcategory_id')->references('id')->on('subcategories');
